[compositor] Allow tracing of composition cycles
