+++ retina-compositor/src/painting.rs
@@ -1,8 +1,6 @@
 // Copyright (C) 2023 Tristan Gerritsen <tristan@thewoosh.org>
 // All Rights Reserved.
 
-use std::sync::OnceLock;
-
 use retina_dom::{
     HtmlElementKind,
     ImageData,
@@ -117,7 +115,6 @@ impl PaintInvoker {
 
     #[instrument(skip_all)]
     pub fn paint(&self, layout_box: &LayoutBox, painter: &mut Painter) {
-        let _guard = CompositorTracingGuard::new();
         self.paint_box(layout_box, painter);
     }
 
@@ -336,42 +333,5 @@ impl PaintInvoker {
                 }
             }
         }
-    }
-}
-
-struct CompositorTracingGuard {
-    _chrome_guard: tracing_chrome::FlushGuard,
-    _tracing_subscriber_guard: tracing::subscriber::DefaultGuard,
-}
-
-static ENABLE_TRACING: OnceLock<bool> = OnceLock::new();
-impl CompositorTracingGuard {
-
-    pub fn new() -> Option<Self> {
-        if !Self::is_enabled() {
-            return None;
-        }
-
-        use tracing_chrome::ChromeLayerBuilder;
-        use tracing_subscriber::prelude::*;
-
-        let (chrome_layer, _chrome_guard) = ChromeLayerBuilder::new().build();
-
-        let _tracing_subscriber_guard = tracing_subscriber::registry()
-            .with(chrome_layer)
-            .set_default();
-
-        Some(Self {
-            _chrome_guard,
-            _tracing_subscriber_guard,
-        })
-    }
-
-    #[inline]
-    pub fn is_enabled() -> bool {
-        *ENABLE_TRACING.get_or_init(|| {
-            std::env::var("RETINA_TRACE")
-                .is_ok_and(|val| val.trim().eq_ignore_ascii_case("1"))
-        })
     }
 }
